pyproject.toml
[tool.poetry]
name = "mini_tw"
version = "0.1.0"
description = ""
authors = ["Matheus Abdias <matheusabdias@b2bit.company>"]

[tool.poetry.scripts]
manage = "manage:main"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.5"
black = "^23.11.0"
pre-commit = "^3.5.0"

[tool.ruff]
src = ["src"]
line-length = 88

exclude = ["*/migrations/*", "manage.py"]

select = [
    "F",
    "I",
    "N",
    "YTT",
    "S",
    "B",
    "COM",
    "C4",
    "DTZ",
    "DJ",
    "EM",
    "EXE",
    "ISC",
    "ICN",
    "G",
    "T20",
    "PYI",
    "Q",
    "RSE",
    "RET",
    "SIM",
    "PL",
    "TRY",
    "UP",
]

ignore = ["DJ001", "S107", "S106", "E501", "S311"]
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
